Add files via upload
You can send mail to no. of persons at the same time as per your
choice , you just have to apply a ‘ for ’ loop in the program.

Basically, its less time consuming becuase you can also fill up the
sender’s email address and password by default in the program , so
u don’t have to enter it everytime u send an mail.

You just have to fill the message and the receiver’s email address ,
and run the program and you gonna go with it and send to how
many receiver’s you want and with subject of the mail too.